dom/base/test/test_bug592829.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=592829
-->
<head>
  <title>Test for Bug 592829</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=592829">Mozilla Bug 592829</a>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 592829 **/

// NOTE! It's imperative that we don't call .init() here. Otherwise we're not
// testing what happens if parsing fails.
// If we ever change how DOMParser initilization works, just update this code
// to create a DOMParser which is not allowed to parse XUL.

var isXUL = true;
try {
  var x =
    SpecialPowers.Cc
      .getService(Components.interfaces.nsIDOMParser)
      .parseFromString('<overlay xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"/>', "text/xml");
  isXUL = x.documentElement.namespaceURI ==
    "http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul";
}
catch (ex) {
  isXUL = false;
}

is(isXUL, false, "We didn't create XUL and we didn't crash!");
</script>
</pre>
</body>
</html>
